refactored irq_handling into own file
